+++ keras_cv_attention_models/coatnet/coatnet.py
@@ -22,7 +22,7 @@ def mhsa_with_multi_head_relative_position_embedding(
 ):
     _, hh, ww, cc = inputs.shape
     key_dim = key_dim if key_dim > 0 else cc // num_heads
-    qk_scale = 1.0 / tf.math.sqrt(tf.cast(key_dim, inputs.dtype))
+    qk_scale = float(1.0 / tf.math.sqrt(tf.cast(key_dim, "float32")))
     out_shape = cc if out_shape is None or not out_weight else out_shape
     qk_out = num_heads * key_dim
     # vv_dim = out_shape // num_heads
